+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where the import menu would load outdated settings values in some cases.
+
 ## 0.4.12 - 2025-10-10
 ### Changed
 * Improved import times when using "Linked Duplicates" as the instance type.